Restructure compact QR coding per cmux file-organization policy
Replace the CmxAttachTicketCompactCoding namespace enum with an
injectable CmxAttachTicketCompactCoder struct, move each compact DTO
(CompactAttachTicket, CompactAttachRoute, CompactAttachEndpoint) into
its own file, and turn the pure private static helpers into file-scope
private functions.
